rename and clean up translators, removes 'pronto'
add a new message type for the high level behavior field from Atlas
