+++ catkin_ws/src/oh_translators/launch/kuka_bridge.launch
