+++ catkin_ws/src/oh_translators/src/lcm2ros_kuka.cpp
