+++ catkin_ws/src/oh_translators/src/ros2lcm_stereo.cpp
