+++ software/drc_lcmtypes/lcmtypes/drc_behavior_t.lcm
@@ -1,0 +1,26 @@
+package drc;
+
+// Current BDI behavior
+// This message was created to not require a dependency on drc_atlas_status_t
+
+struct behavior_t
+{
+  int64_t utime;
+  int32_t behavior;
+
+  // Behavior - to remain synced with drc_atlas_status_t:
+  const int32_t BEHAVIOR_NONE=0;        //!< Robot currently not in any behavior; software freeze, controls disabled.
+  const int32_t BEHAVIOR_FREEZE=1;      //!< All joints frozen in current position.
+  const int32_t BEHAVIOR_STAND_PREP=2;  //!< Robot assumes a balanced posture, preparing for Stand.
+  const int32_t BEHAVIOR_STAND=3;       //!< Statically stable stand.
+  const int32_t BEHAVIOR_WALK=4;        //!< Dynamically stable walk.
+  const int32_t BEHAVIOR_STEP=5;        //!< Statically stable, single-step walk.
+  const int32_t BEHAVIOR_MANIPULATE=6;  //!< Statically stable stand, with upper body joints controlled by user.
+  const int32_t BEHAVIOR_USER=7;        //!< Setpoints and gains set by user.
+  const int32_t BEHAVIOR_CALIBRATE=8;   //!< Robot moves through various calibration motions, returns to Freeze.
+  const int32_t BEHAVIOR_SOFT_STOP=9;   //!< "Safe" falling behavior, robot tucks its arms and legs and uses soft position gains
+}
+
+
+
+
