+++ software/drc_lcmtypes/lcmtypes/drc_map_snapshot_request_t.lcm
